requirements.txt
spotipy==2.23.0
numpy==1.24.2
scikit-learn==1.2.2
pandas==2.0.0
pyinstaller==5.10.1
google-cloud-secret-manager==2.10.0
Flask==2.2.3
requests==2.28.2